[frontend] Refactor AppIntProvider.tsx
