+++ opencti-platform/opencti-front/lang/es.json
@@ -2434,5 +2434,7 @@
   "Task": "Tarea",
   "Do you want to unlink this task ?": "¿Desea desvincular esta tarea?",
   "CREATE": "CREAR",
-  "entity_Analyses": "entity_Analyses"
+  "entity_Analyses": "entity_Analyses",
+  "There are no opinions. You should create some before.": "No hay opiniones. Deberías crear algunas antes.",
+  "workflow_id": "Estado"
 }
