--- opencti-platform/opencti-front/src/utils/Localization.js
@@ -1,17 +1,0 @@
-import messages_en from '../../lang/en.json';
-import messages_fr from '../../lang/fr.json';
-import messages_es from '../../lang/es.json';
-import messages_zh from '../../lang/zh.json';
-import messages_ja from '../../lang/ja.json';
-
-const i18n = {
-  messages: {
-    'es-es': messages_es,
-    'fr-fr': messages_fr,
-    'ja-jp': messages_ja,
-    'zh-cn': messages_zh,
-    'en-us': messages_en,
-  },
-};
-
-export default i18n;
